[AIRFLOW-5790] Move AWS Hook to /providers/aws
