+++ docs/operators-and-hooks-ref.rst
@@ -289,7 +289,7 @@ AWS: Amazon Web Services
 
 Airflow has support for `Amazon Web Services <https://aws.amazon.com/>`__.
 
-All hooks are based on :mod:`airflow.contrib.hooks.aws_hook`.
+All hooks are based on :mod:`airflow.providers.aws.hooks.aws`.
 
 Service operators and hooks
 '''''''''''''''''''''''''''
